requirements.txt
# Core dependencies for LLM Council MVP
openai>=1.30.0
pydantic>=2.0.0
pyyaml>=6.0
aiohttp>=3.8.0
click>=8.0.0

# Development dependencies  
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.0.0

# Optional framework for v2 migration
# crewai>=0.28.0
# langgraph>=0.0.40
